wip/test: scan and total for basic prices

+++ Checkout/ICheckout.cs
@@ -3,6 +3,6 @@
     interface ICheckout
     {
         void Scan(string item);
-        int GetTotalPrice();
+        float GetTotalPrice();
     }
 }
